+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file. For commit 
 
  **BugFixes**:
  - root download of a shared directory returns HTTP 500 (#2807) (#2810) (#2821) (#2818)
+ - OnlyOffice is inaccessible if share has optional password (#2811)
+
 ## v1.5.2-stable
 
  **Security**: